@@ -7,12 +7,16 @@ import com.htec.filesystem.entity.FolderEntity;
 import com.htec.filesystem.entity.ShelfEntity;
 import com.htec.filesystem.exception.ExceptionSupplier;
 import com.htec.filesystem.mapper.ShelfItemMapper;
+import com.htec.filesystem.model.request.LogRequestModel;
 import com.htec.filesystem.model.request.RenameFileRequestModel;
 import com.htec.filesystem.model.response.FileResponseModel;
 import com.htec.filesystem.repository.FileRepository;
 import com.htec.filesystem.repository.FolderRepository;
 import com.htec.filesystem.repository.ShelfRepository;
 import com.htec.filesystem.util.FileUtil;
+import com.htec.filesystem.util.FunctionEvents;
+import com.htec.filesystem.validator.FileSystemValidator;
+import org.springframework.core.io.FileSystemResource;
 import org.springframework.data.util.Pair;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -20,8 +24,8 @@ import org.springframework.util.StreamUtils;
 
 import java.io.File;
 import java.io.FileInputStream;
-import java.io.FileOutputStream;
 import java.io.IOException;
+import java.io.OutputStream;
 import java.nio.file.*;
 import java.time.LocalDateTime;
 import java.util.*;
@@ -42,13 +46,19 @@ public class FileService {
     private final FileRepository fileRepository;
     private final FolderRepository folderRepository;
     private final ShelfRepository shelfRepository;
+    private final FileSystemValidator fileSystemValidator;
 
-    public FileService(UserAPICallService userAPICallService, FileRepository fileRepository, FolderRepository folderRepository, ShelfRepository shelfRepository) {
+    public FileService(UserAPICallService userAPICallService,
+                       FileRepository fileRepository,
+                       FolderRepository folderRepository,
+                       ShelfRepository shelfRepository,
+                       FileSystemValidator fileSystemValidator) {
 
         this.userAPICallService = userAPICallService;
         this.fileRepository = fileRepository;
         this.folderRepository = folderRepository;
         this.shelfRepository = shelfRepository;
+        this.fileSystemValidator = fileSystemValidator;
     }
 
     public void saveUserProfilePicture(Long id, Map<String, Pair<String, String>> files) {
@@ -97,7 +107,9 @@ public class FileService {
         return new FileResponseModel(imageBytes, path);
     }
 
-    public void saveFile(Long shelfId, Long folderId, Map<String, Pair<String, String>> files, Long userId) {
+    public List<Long> saveFile(Long shelfId, Long folderId, Map<String, Pair<String, String>> files, Long userId) {
+
+        List<Long> newFileIds = new ArrayList<>();
 
         if (files == null) {
             throw ExceptionSupplier.couldNotUploadFile.get();
@@ -120,7 +132,6 @@ public class FileService {
             throw ExceptionSupplier.userNotAllowedToAccessShelf.get();
         }
 
-
         for (Map.Entry<String, Pair<String, String>> filesPair : files.entrySet()) {
 
             byte[] bytes = Base64.getDecoder().decode(filesPair.getValue().getSecond());
@@ -140,31 +151,61 @@ public class FileService {
                 dbPath = folderEntity.getPath() + pathSeparator + fileName;
                 localPath = userPath + folderEntity.getPath() + pathSeparator;
 
+                int fileCounter = 0;
                 if (fileRepository.findByNameAndParentFolderId(fileName, folderId).isPresent()) {
-                    throw ExceptionSupplier.fileAlreadyExists.get();
+                    String newFileName = fileName;
+
+                    while (fileRepository.findByNameAndParentFolderId(newFileName, folderId).isPresent()) {
+                        fileCounter++;
+                        int extensionIndex = fileName.lastIndexOf('.');
+                        if (extensionIndex != -1) {
+
+                            String nameWithoutExtension = fileName.substring(0, extensionIndex);
+                            String extension = fileName.substring(extensionIndex);
+                            newFileName = nameWithoutExtension + "(" + fileCounter + ")" + extension;
+                        } else {
+                            newFileName = fileName + "(" + fileCounter + ")";
+                        }
+                    }
+                    fileName = newFileName;
                 }
 
             } else {
 
+                int fileCounter = 0;
+                if (fileRepository.findByNameAndShelfIdAndParentFolderIdIsNull(fileName, shelfId).isPresent()) {
+                    String newFileName = fileName;
+
+                    while (fileRepository.findByNameAndShelfIdAndParentFolderIdIsNull(newFileName, shelfId).isPresent()) {
+
+                        newFileName = fileName;
+                        fileCounter++;
+                        int extensionIndex = fileName.lastIndexOf('.');
+                        if (extensionIndex != -1) {
+
+                            String nameWithoutExtension = fileName.substring(0, extensionIndex);
+                            String extension = fileName.substring(extensionIndex);
+                            newFileName = nameWithoutExtension + "(" + fileCounter + ")" + extension;
+                        } else {
+                            newFileName = fileName + "(" + fileCounter + ")";
+                        }
+                    }
+                    fileName = newFileName;
+                }
                 localPath = userPath + shelfEntity.getUserId() + pathSeparator + "shelves" + pathSeparator + shelfId + pathSeparator;
                 dbPath = shelfEntity.getUserId() + pathSeparator + "shelves" + pathSeparator + shelfId + pathSeparator + fileName;
-
-                if (fileRepository.findByNameAndShelfIdAndParentFolderIdIsNull(fileName, shelfId).isPresent()) {
-                    throw ExceptionSupplier.fileAlreadyExists.get();
-                }
-            }
-
-            if (fileRepository.findByPath(dbPath).isPresent()) {
-                throw ExceptionSupplier.fileAlreadyExists.get();
             }
 
             String uploadDir = homePath + localPath;
             FileUtil.saveFile(uploadDir, fileName, bytes);
-            saveFileIntoDB(dbPath, fileName, shelfId, folderId);
+
+            newFileIds.add(saveFileIntoDB(dbPath, fileName, shelfId, folderId));
         }
+
+        return newFileIds;
     }
 
-    public void saveFileIntoDB(String filePath, String fileName, long shelfId, long folderId) {
+    public Long saveFileIntoDB(String filePath, String fileName, long shelfId, long folderId) {
 
         FileEntity fileEntity = new FileEntity();
 
@@ -175,7 +216,9 @@ public class FileService {
         if (folderId != 0) fileEntity.setParentFolderId(folderId);
 
         fileEntity.setCreatedAt(LocalDateTime.now());
-        fileRepository.save(fileEntity);
+        fileRepository.saveAndFlush(fileEntity);
+
+        return fileEntity.getId();
     }
 
     public void copyFile(Long fileId, Long shelfId, Long userId) {
@@ -183,12 +226,12 @@ public class FileService {
         List<ShelfEntity> shelfEntities = shelfRepository.findAllByUserId(userId);
 
         FileEntity fileEntity = fileRepository.findById(fileId)
-                .orElseThrow(ExceptionSupplier.noFileWithGivenId);
+                .orElseThrow(ExceptionSupplier.fileNotFound);
 
         shelfRepository.findById(shelfId)
-                .orElseThrow(ExceptionSupplier.noShelfWithGivenId);
+                .orElseThrow(ExceptionSupplier.shelfNotFound);
 
-        if(!shelfEntities.stream().map(ShelfEntity::getId).collect((Collectors.toList())).contains(fileEntity.getShelfId())) {
+        if (!shelfEntities.stream().map(ShelfEntity::getId).collect((Collectors.toList())).contains(fileEntity.getShelfId())) {
             throw ExceptionSupplier.userNotAllowedToAccessShelf.get();
         }
 
@@ -208,7 +251,7 @@ public class FileService {
         fileRepository.save(backupFile);
 
         try {
-            Files.copy(Paths.get(src),Paths.get(dst), StandardCopyOption.REPLACE_EXISTING);
+            Files.copy(Paths.get(src), Paths.get(dst), StandardCopyOption.REPLACE_EXISTING);
         } catch (IOException e) {
             throw ExceptionSupplier.couldNotCopyFile.get();
         }
@@ -329,7 +372,6 @@ public class FileService {
         }
     }
 
-
     private void moveFilesToTrash(List<FileEntity> fileEntities, Boolean trashVisible) {
 
         for (FileEntity fileEntity : fileEntities) {
@@ -381,10 +423,6 @@ public class FileService {
         }
     }
 
-    private void recoverFilesInDb() {
-
-    }
-
     public void fileRename(Long userId, RenameFileRequestModel renameFileRequestModel) {
 
         String fileName = renameFileRequestModel.getFileName();
@@ -397,6 +435,8 @@ public class FileService {
         ShelfEntity shelfEntity = shelfRepository.findById(fileEntity.getShelfId())
                 .orElseThrow(ExceptionSupplier.noShelfWithGivenId);
 
+        fileSystemValidator.isFileNameValid(renameFileRequestModel.getFileName());
+
         if (!Objects.equals(shelfEntity.getUserId(), userId))
             throw ExceptionSupplier.userNotAllowedToAccessFile.get();
 
@@ -406,9 +446,12 @@ public class FileService {
             throw ExceptionSupplier.fileAlreadyExists.get();
 
         int dotIndex = fileEntity.getName().lastIndexOf(".");
-        String fileExtension = fileEntity.getName().substring(dotIndex);
 
-        fileName += fileExtension;
+        if (dotIndex != -1) {
+
+            String fileExtension = fileEntity.getName().substring(dotIndex);
+            fileName += fileExtension;
+        }
 
         String oldFilePath = homePath + userPath + fileEntity.getPath();
         File oldFile = new File(oldFilePath);
@@ -435,7 +478,7 @@ public class FileService {
         return new ArrayList<>(ShelfItemMapper.INSTANCE.fileEntitiesToShelfItemDTOs(fileEntities));
     }
 
-    public void deleteFile(Long userId, List<Long> fileIds) {
+    public Long deleteFile(Long userId, List<Long> fileIds) {
 
         List<FileEntity> fileEntities = fileRepository.findAllByUserIdAndDeletedAndIdIn(userId, true, fileIds);
 
@@ -447,6 +490,8 @@ public class FileService {
             throw ExceptionSupplier.userNotAllowedToDeleteFile.get();
         }
 
+        Long shelfId = fileEntities.isEmpty() ? null : fileEntities.get(0).getShelfId();
+
         for (FileEntity fileEntity : fileEntities) {
             String fullPath = homePath + userPath + fileEntity.getPath();
             if (!(new File(fullPath)).delete()) {
@@ -455,11 +500,37 @@ public class FileService {
         }
 
         fileRepository.deleteAll(fileEntities);
+        return shelfId;
     }
 
-    public void downloadFilesToZip(AuthUser user, List<Long> fileIds) {
+    public ZipOutputStream downloadFilesToZip(List<Long> fileIds, OutputStream outputStream) {
 
-        List<FileEntity> fileEntities = fileRepository.findAllByUserIdAndDeletedAndIdIn(user.getId(), false, fileIds);
+        List<FileEntity> fileEntities = fileRepository.findAllByUserIdAndDeletedAndIdIn(8L, false, fileIds);
+
+        validation(fileIds, fileEntities);
+
+        try (ZipOutputStream zippedOut = new ZipOutputStream(outputStream)) {
+
+            for (FileEntity fileEntity : fileEntities) {
+
+                String fullPath = homePath + userPath + fileEntity.getPath();
+                FileSystemResource resource = new FileSystemResource(fullPath);
+                ZipEntry entry = new ZipEntry(Objects.requireNonNull(resource.getFilename()));
+                entry.setSize(resource.contentLength());
+                entry.setTime(System.currentTimeMillis());
+                zippedOut.putNextEntry(entry);
+                StreamUtils.copy(resource.getInputStream(), zippedOut);
+                zippedOut.closeEntry();
+            }
+
+            zippedOut.finish();
+            return zippedOut;
+        } catch (IOException e) {
+            throw ExceptionSupplier.couldNotDownloadFiles.get();
+        }
+    }
+
+    private void validation(List<Long> fileIds, List<FileEntity> fileEntities) {
 
         if (fileEntities.size() != fileIds.size()) {
             throw ExceptionSupplier.filesNotFound.get();
@@ -468,27 +539,62 @@ public class FileService {
         if (!fileEntities.stream().map(FileEntity::getId).collect(Collectors.toList()).containsAll(fileIds)) {
             throw ExceptionSupplier.userNotAllowedToDownloadFile.get();
         }
+    }
 
-        try (FileOutputStream fos = new FileOutputStream("file-system/src/main/resources/multiCompressed.zip"); ZipOutputStream zipOut = new ZipOutputStream(fos)) {
+    public void logFile(Long backupFileId, LogRequestModel logRequestModel) {
 
-            for (FileEntity fileEntity : fileEntities) {
+        Integer eventId = logRequestModel.getEventId();
+        String fileName = logRequestModel.getFileName();
 
-                String fullPath = homePath + userPath + fileEntity.getPath();
-                File fileToZip = new File(fullPath);
-                try (FileInputStream fis = new FileInputStream(fileToZip)) {
-                    ZipEntry zipEntry = new ZipEntry(fileToZip.getName());
-                    zipOut.putNextEntry(zipEntry);
+        FileEntity logFile = fileRepository.findById(backupFileId)
+                .orElseThrow(ExceptionSupplier.fileNotFound);
 
-                    byte[] bytes = new byte[1024];
-                    int length;
-                    while ((length = fis.read(bytes)) >= 0) {
-                        zipOut.write(bytes, 0, length);
-                    }
-                }
-            }
-            zipOut.finish();
+        FunctionEvents event = FunctionEvents.values()[eventId - 1];
+
+        String logMessage = event.getEvent() + "\t" + fileName + "\t" + LocalDateTime.now() + "\n";
+
+        String sourceFilePath = homePath + userPath + logFile.getPath();
+
+        try {
+            Files.writeString(Path.of(sourceFilePath), logMessage, StandardOpenOption.APPEND);
         } catch (IOException e) {
-            throw ExceptionSupplier.couldNotDownloadFiles.get();
+            e.printStackTrace();
         }
+    }
+
+    public Long createLogFile(Long shelfId, Long folderId, String logFileName, Long userId) {
+
+        Map<String, Pair<String, String>> files = new HashMap<>();
+
+        Pair<String, String> file = Pair.of(logFileName, "");
+        files.put(logFileName, file);
+
+        List<Long> logFileIds = saveFile(shelfId, folderId, files, userId);
+
+        return logFileIds.get(0);
+    }
+
+    public Long getLogFileId(Long shelfId, String logFileName, Long userId) {
+
+        Optional<FileEntity> fileEntityOptional = fileRepository.findByNameAndShelfIdAndParentFolderIdIsNull(logFileName, shelfId);
+        if (fileEntityOptional.isPresent()) {
+
+            return fileEntityOptional.get().getId();
+        }
+
+        return createLogFile(shelfId, 0L, logFileName, userId);
+    }
+
+    public Map<Long, String> getFileNamesFromIds(List<Long> fileIds) {
+
+        Map<Long, String> filesToDelete = new HashMap<>();
+
+        List<FileEntity> fileEntities = fileRepository.findAllByIdIn(fileIds);
+
+        for (FileEntity fileEntity : fileEntities) {
+            filesToDelete.put(fileEntity.getId(), fileEntity.getRealName());
+        }
+
+        return filesToDelete;
     }
 }
