@@ -2,18 +2,23 @@ package com.htec.filesystem.controller;
 
 import com.htec.filesystem.annotation.AuthUser;
 import com.htec.filesystem.annotation.AuthenticationUser;
+import com.htec.filesystem.model.request.LogRequestModel;
 import com.htec.filesystem.model.request.RenameFileRequestModel;
 import com.htec.filesystem.model.response.FileResponseModel;
 import com.htec.filesystem.model.response.TextResponseMessage;
+import com.htec.filesystem.service.EventService;
 import com.htec.filesystem.service.FileService;
+import com.htec.filesystem.util.FunctionEvents;
 import org.springframework.data.util.Pair;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.servlet.mvc.method.annotation.StreamingResponseBody;
 
 import java.io.IOException;
+import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 
@@ -22,9 +27,11 @@ import java.util.Map;
 public class FileController {
 
     private final FileService fileService;
+    private final EventService eventService;
 
     private final String FILE_UPLOADED = "File Uploaded";
     private final String FILE_COPIED = "File Copied";
+    private final String FILE_LOGGED = "File logged";
     private final String FILE_DOWNLOADED = "File Downloaded";
     private final String FILE_RENAMED = "File renamed";
     private final String IMAGE_UPLOADED = "Image Uploaded";
@@ -32,14 +39,15 @@ public class FileController {
     private final String FILES_RECOVERED_FROM_TRASH = "File/s recovered from trash.";
     private final String FILES_DELETED = "File/s deleted.";
 
-    public FileController(FileService fileService) {
+    public FileController(FileService fileService,
+                          EventService eventService) {
         this.fileService = fileService;
+        this.eventService = eventService;
     }
 
 
     @PostMapping("/upload/profile/{id}")
-    public ResponseEntity<TextResponseMessage> uploadProfilePicture(@RequestBody Map<String, Pair<String, String>> files,
-                                                                    @PathVariable Long id) {
+    public ResponseEntity<TextResponseMessage> uploadProfilePicture(@RequestBody Map<String, Pair<String, String>> files, @PathVariable Long id) {
 
         fileService.saveUserProfilePicture(id, files);
         return ResponseEntity.status(HttpStatus.OK).body(new TextResponseMessage(IMAGE_UPLOADED, HttpStatus.OK.value()));
@@ -51,16 +59,18 @@ public class FileController {
                                           @RequestParam(value = "file") boolean file) {
 
         FileResponseModel fileResponseModel = fileService.getFile(user.getId(), id, file);
+
+        eventService.reportEvent(FunctionEvents.DOWNLOAD, Collections.singletonList(id), user.getId(), null, null);
+
         return ResponseEntity.ok().contentType(MediaType.MULTIPART_FORM_DATA).body(fileResponseModel.getImageContent());
     }
 
-    @PostMapping(value = "/zip-download", produces = "application/zip")
-    public ResponseEntity zipDownload(@AuthenticationUser AuthUser user, @RequestBody List<Long> fileIds) {
+    @GetMapping(value = "/zip-download")
+    public ResponseEntity<StreamingResponseBody> zipDownload(@RequestParam List<Long> fileIds) {
 
-        fileService.downloadFilesToZip(user, fileIds);
         return ResponseEntity.ok()
-                .contentType(MediaType.parseMediaType("application/zip"))
-                .header(HttpHeaders.CONTENT_DISPOSITION, "attachment; filename\"" + "asde" + "\"").build();
+                .header(HttpHeaders.CONTENT_DISPOSITION, "attachment; filename=export.zip")
+                .body(outputStream -> fileService.downloadFilesToZip(fileIds, outputStream));
     }
 
     @GetMapping("/preview/{id}")
@@ -77,8 +87,21 @@ public class FileController {
                                                           @PathVariable Long folderId,
                                                           @AuthenticationUser AuthUser authUser) {
 
-        fileService.saveFile(shelfId, folderId, files, authUser.getId());
+        List<Long> newFileIds = fileService.saveFile(shelfId, folderId, files, authUser.getId());
+
+        eventService.reportEvent(FunctionEvents.UPLOAD, newFileIds, authUser.getId(), shelfId, null);
+
         return ResponseEntity.status(HttpStatus.OK).body(new TextResponseMessage(FILE_UPLOADED, HttpStatus.OK.value()));
+    }
+
+    @PostMapping("/log/{shelfId}")
+    public ResponseEntity<Long> getLogFileId(@RequestBody String logFileName,
+                                             @PathVariable Long shelfId,
+                                             @AuthenticationUser AuthUser authUser) {
+
+        Long logFileId = fileService.getLogFileId(shelfId, logFileName, authUser.getId());
+
+        return ResponseEntity.status(HttpStatus.OK).body(logFileId);
     }
 
     @PostMapping("/copy/{fileId}/shelf/{shelfId}/user/{userId}")
@@ -88,6 +111,14 @@ public class FileController {
 
         fileService.copyFile(fileId, shelfId, userId);
         return ResponseEntity.status(HttpStatus.OK).body(new TextResponseMessage(FILE_COPIED, HttpStatus.OK.value()));
+    }
+
+    @PutMapping("/log/event/backupFileId/{backupFileId}")
+    public ResponseEntity<TextResponseMessage> logFile(@PathVariable Long backupFileId,
+                                                       @RequestBody LogRequestModel logRequestModel) {
+
+        fileService.logFile(backupFileId, logRequestModel);
+        return ResponseEntity.status(HttpStatus.OK).body(new TextResponseMessage(FILE_LOGGED, HttpStatus.OK.value()));
     }
 
     @PutMapping("/move-to-trash")
@@ -106,16 +137,20 @@ public class FileController {
     }
 
     @DeleteMapping
-    public ResponseEntity<TextResponseMessage> deleteFile(@AuthenticationUser AuthUser user, @RequestBody List<Long> fileIds) throws IOException {
+    public ResponseEntity<TextResponseMessage> deleteFile(@AuthenticationUser AuthUser user, @RequestBody List<Long> fileIds) {
 
-        fileService.deleteFile(user.getId(), fileIds);
+        Map<Long, String> filesToDelete = fileService.getFileNamesFromIds(fileIds);
+        Long deletedFilesShelfId = fileService.deleteFile(user.getId(), fileIds);
+
+        eventService.reportEvent(FunctionEvents.DELETE, fileIds, user.getId(), deletedFilesShelfId, filesToDelete);
+
         return ResponseEntity.ok().body(new TextResponseMessage(FILES_DELETED, HttpStatus.OK.value()));
     }
 
     @PutMapping("/recover")
     public ResponseEntity<TextResponseMessage> recoverFile(@AuthenticationUser AuthUser user, @RequestBody List<Long> fileIds) {
 
-        fileService.updateDeletedFiles(user, fileIds , false, true);
+        fileService.updateDeletedFiles(user, fileIds, false, true);
         return ResponseEntity.ok().body(new TextResponseMessage(FILES_RECOVERED_FROM_TRASH, HttpStatus.OK.value()));
     }
 }
