@@ -5,7 +5,6 @@ import { FaRegTimesCircle } from 'react-icons/fa';
 import { toast } from 'react-toastify';
 import { FunctionForm, FunctionFormInner } from '../form/form-styles';
 import { InputField } from '../UI/input/InputField';
-import { ModalButtonDivider } from '../layout/layout.styles';
 import { Button } from '../UI/button';
 import functionService from '../../services/functionService';
 import { Error, H2 } from '../text/text-styles';
@@ -21,6 +20,7 @@ import {
   RadioInner,
   RadioSubTitle,
   RadioTitle,
+  FunctionModalFooter,
 } from './modal.styles';
 import { theme } from '../../theme';
 import { InputSubTitle, InputTitle } from '../UI/input/input-styles';
@@ -33,7 +33,7 @@ import {
   languageOptions,
   logTrigger,
 } from '../../utils/fixtures/functionOptions';
-import { useAppSelector } from '../../store/hooks';
+import { config } from '../../utils/validation/config/functionModalValidations';
 
 const FunctionModal = ({
   onCloseModal,
@@ -44,6 +44,7 @@ const FunctionModal = ({
     register,
     handleSubmit,
     reset,
+    setValue,
     watch,
     formState: { errors },
   } = useForm();
@@ -51,18 +52,20 @@ const FunctionModal = ({
   const [optionValue, setOptionValue] = useState<string>('scratch');
 
   const handleCloseModal = () => {
-    onCloseModal(false);
+    onCloseModal();
   };
-  const isLoading = useAppSelector((state) => state.loading.loading);
 
   const onSubmit = (formData: FunctionFormData) => {
+    const newFormData: FunctionFormData = formData;
+    newFormData.name = newFormData.name.trim();
+
     if (formData.function === 'backup') {
       formData.eventId = 1;
       functionService
         .createPredefinedFunction(formData)
         .then(() => {
           toast.success('Function created');
-          onCloseModal(false);
+          onCloseModal();
           onGetData();
         })
         .catch((err) => toast.error(err.response?.data.message));
@@ -71,7 +74,7 @@ const FunctionModal = ({
         .createCustomfunction(formData)
         .then(() => {
           toast.success('Function created');
-          onCloseModal(false);
+          onCloseModal();
           onGetData();
         })
         .catch((err) => {
@@ -80,24 +83,12 @@ const FunctionModal = ({
     }
   };
 
-  const validations = {
-    required: 'This field is required',
-    maxLength: {
-      value: 50,
-      message: 'Function name can not be longer than 50 characters',
-    },
-  };
+  const predefinedFunction = watch('function');
+  const validations = config(watch);
 
   const handleChange = ({ target }: React.ChangeEvent<HTMLInputElement>) => {
     setOptionValue(target.value);
     reset();
-  };
-
-  const predefinedFunction = watch('function');
-
-  const validateBackupShelf = {
-    validate: (value: string) =>
-      value !== watch('shelfId') || "Can't select same shelfs",
   };
 
   const modal = (
@@ -150,23 +141,22 @@ const FunctionModal = ({
                     preset for common use cases.
                   </RadioSubTitle>
                 </div>
-
-                <input
-                  type="radio"
-                  id="blueprint"
-                  name="code"
-                  value="blueprint"
-                  checked={optionValue === 'blueprint'}
-                  onChange={handleChange}
-                />
               </RadioInner>
+              <input
+                type="radio"
+                id="blueprint"
+                name="code"
+                value="blueprint"
+                checked={optionValue === 'blueprint'}
+                onChange={handleChange}
+              />
               <img
                 src={`${process.env.PUBLIC_URL}/assets/images/coding.png`}
                 alt="code symbol"
               />
             </RadioLabel>
           </RadioContainer>
-          <FunctionForm onSubmit={handleSubmit(onSubmit)}>
+          <FunctionForm onSubmit={handleSubmit(onSubmit)} id="function-form">
             <FunctionFormInner>
               {optionValue === 'scratch' ? (
                 <>
@@ -177,7 +167,7 @@ const FunctionModal = ({
                   <InputField
                     placeholder="Function name"
                     error={errors.name}
-                    {...register('name', validations)}
+                    {...register('name', validations.functionValidation)}
                   />
                   <Error>{errors?.functionName?.message}</Error>
                   <InputTitle>Runtime</InputTitle>
@@ -186,10 +176,12 @@ const FunctionModal = ({
                   </InputSubTitle>
                   <Select
                     optionsData={languageOptions}
+                    register={register}
                     selectName="language"
                     error={errors.language}
-                    register={register}
                     placeHolder="Select language"
+                    setValue={setValue}
+                    validation={validations.basicValidation}
                   />
                   <InputTitle>Select event trigger</InputTitle>
                   <InputSubTitle>
@@ -199,8 +191,10 @@ const FunctionModal = ({
                     optionsData={eventTriggerOptions}
                     register={register}
                     selectName="eventId"
-                    error={errors.trigger}
+                    error={errors.eventId}
                     placeHolder="Select function trigger"
+                    setValue={setValue}
+                    validation={validations.basicValidation}
                   />
                   <InputTitle>Bind function to shelf</InputTitle>
                   <InputSubTitle>
@@ -212,6 +206,8 @@ const FunctionModal = ({
                     selectName="shelfId"
                     error={errors.shelfId}
                     placeHolder="Select shelf"
+                    setValue={setValue}
+                    validation={validations.basicValidation}
                   />
                 </>
               ) : (
@@ -226,6 +222,8 @@ const FunctionModal = ({
                     selectName="function"
                     error={errors.function}
                     placeHolder="Select function"
+                    setValue={setValue}
+                    validation={validations.basicValidation}
                   />
                   {predefinedFunction && (
                     <>
@@ -236,7 +234,7 @@ const FunctionModal = ({
                       <InputField
                         placeholder="Function name"
                         error={errors.name}
-                        {...register('name', validations)}
+                        {...register('name', validations.functionValidation)}
                       />
                       <InputTitle>Bind function to shelf</InputTitle>
                       <InputSubTitle>
@@ -248,6 +246,8 @@ const FunctionModal = ({
                         selectName="shelfId"
                         error={errors.shelfId}
                         placeHolder="Select shelf"
+                        setValue={setValue}
+                        validation={validations.basicValidation}
                       />
                     </>
                   )}
@@ -260,10 +260,11 @@ const FunctionModal = ({
                       <Select
                         optionsData={data}
                         register={register}
-                        validation={validateBackupShelf}
                         selectName="functionParam"
                         error={errors.functionParam}
                         placeHolder="Select backup shelf"
+                        setValue={setValue}
+                        validation={validations.backupShelf}
                       />
                     </>
                   )}
@@ -276,29 +277,31 @@ const FunctionModal = ({
                       <Select
                         optionsData={logTrigger}
                         register={register}
-                        selectName="trigger"
-                        error={errors.trigger}
+                        selectName="eventId"
+                        error={errors.eventId}
+                        setValue={setValue}
                         placeHolder="Select function trigger"
+                        validation={validations.basicValidation}
                       />
                     </>
                   )}
                 </>
               )}
             </FunctionFormInner>
-            <ModalButtonDivider>
-              <Button
-                variant="lightBordered"
-                onClick={handleCloseModal}
-                size="large"
-              >
-                Cancel
-              </Button>
-              <Button spinner isLoading={isLoading} size="large">
-                Submit
-              </Button>
-            </ModalButtonDivider>
           </FunctionForm>
         </FunctionContainer>
+        <FunctionModalFooter>
+          <Button
+            variant="lightBordered"
+            onClick={handleCloseModal}
+            size="large"
+          >
+            Cancel
+          </Button>
+          <Button size="large" type="submit" form="function-form">
+            Submit
+          </Button>
+        </FunctionModalFooter>
       </FunctionModalContainer>
     </Backdrop>
   );
