@@ -18,12 +18,13 @@ export const Backdrop = styled.div`
   z-index: 9999;
 `;
 
-export const ModalContainer = styled.div`
+export const ModalContainer = styled.div<ModalStyleProps>`
   min-width: 300px;
   max-width: 500px;
   position: absolute;
 
-  background-color: ${theme.colors.white};
+  background-color: ${({ background }) => background || theme.colors.white};
+  color: ${({ color }) => color || theme.colors.primary};
   border-radius: 10px;
 
   display: flex;
@@ -39,9 +40,7 @@ export const Header = styled.div`
 
   max-height: 70px;
   padding: ${theme.space.none} ${theme.space.sm};
-
   width: 100%;
-  color: ${theme.colors.primary};
 `;
 
 export const HeaderItem = styled.div`
@@ -59,8 +58,6 @@ export const Body = styled.div`
   flex-direction: column;
   justify-content: center;
   align-items: center;
-
-  color: ${theme.colors.primary};
 `;
 
 export const Footer = styled.div`
@@ -72,7 +69,7 @@ export const Footer = styled.div`
   bottom: 0;
   height: 90px;
   width: 100%;
-  border-top: 1px solid ${theme.colors.primary};
+  border-top: 1px solid ${({ color }) => color || theme.colors.primary};
 `;
 
 export const Close = styled.button<ButtonProps>`
@@ -89,8 +86,10 @@ export const DeleteModalBody = styled.div`
 
 export const RadioContainer = styled.div`
   display: flex;
+  text-align: justify;
+  line-height: 20px;
 
-  @media (max-width: ${theme.breakpoints.md}) {
+  @media (max-width: ${theme.breakpoints.sm}) {
     flex-direction: column;
   }
 `;
@@ -103,6 +102,7 @@ export const RadioLabel = styled.label<FunctionStyleProps>`
   border: 1px;
   width: 100%;
   height: 200px;
+  position: relative;
   margin-right: ${theme.space.lg};
   &:last-of-type {
     margin-right: 0;
@@ -114,10 +114,14 @@ export const RadioLabel = styled.label<FunctionStyleProps>`
   }
   input {
     cursor: pointer;
+    position: absolute;
+    top: 15px;
+    right: 15px;
   }
 
   @media (max-width: ${theme.breakpoints.md}) {
     margin-bottom: ${theme.space.sm};
+    padding: 0 20px;
   }
 `;
 
@@ -126,11 +130,15 @@ export const RadioInner = styled.div`
   display: flex;
   justify-content: space-between;
   min-height: 120px;
+
+  @media (max-width: ${theme.breakpoints.md}) {
+    text-align: left;
+  }
 `;
 export const FunctionModalContainer = styled.div`
   background: ${theme.colors.secondary};
-
   height: 700px;
+  max-width: 700px;
   color: white;
   display: flex;
   flex-direction: column;
@@ -138,13 +146,21 @@ export const FunctionModalContainer = styled.div`
   border-radius: 10px;
   margin: ${theme.space.md};
 `;
+
 export const FunctionContainer = styled.div`
   overflow-y: auto;
   padding: 0px ${theme.space.lg} ${theme.space.sm};
   border-radius: 55px;
   height: 100%;
 `;
-export const RadioTitle = styled.p``;
+export const RadioTitle = styled.p`
+  margin: 0;
+  margin-bottom: ${theme.space.sm};
+
+  @media (max-width: ${theme.breakpoints.md}) {
+    margin: ${theme.space.sm} 0;
+  }
+`;
 export const RadioSubTitle = styled.p`
   margin: 0;
   font-size: ${theme.fontSizes.sm};
@@ -154,9 +170,19 @@ export const FunctionHeader = styled.div`
   display: flex;
   align-items: center;
   justify-content: space-between;
-
   height: 70px;
   padding: ${theme.space.none} ${theme.space.sm};
-
   width: 100%;
+`;
+
+export const FunctionModalFooter = styled.div`
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding: ${theme.space.md} ${theme.space.lg};
+  width: 100%;
+
+  @media (max-width: ${theme.breakpoints.sm}) {
+    flex-direction: column;
+  }
 `;
