@@ -1,4 +1,3 @@
-import { Dispatch, SetStateAction } from 'react';
 import { RegisterOptions } from 'react-hook-form';
 import { Role } from '../utils/enums/roles';
 import { SortingDirection } from '../utils/enums/table';
@@ -269,5 +268,13 @@ export type FunctionFormData =
   | BackupFunctionData
   | LogFunctionData;
 
-
-
+export interface FunctionData {
+  code: string;
+  eventId: number;
+  eventName: string;
+  functionId: number;
+  functionName: string;
+  language: string;
+  shelfId: number;
+  shelfName: string;
+}
